Quantum Solution to the Discrete Logarithm Problem
This project aims to get a solution to the discrete logarithmic (Dlog) problem by the quantum computer algorithm. The solution is base one finding the half-bit (HB) values of the power using an oracle-based quantum algorithm. The oracle function is repeatedly invoked to calculate the HB of each bit in the power of the Dlog of a given base and modulus. The proposed Dlog algorithm works for any group of order 2k and calculate a potential value of the power due to the error associated with the oracle-based quantum algorithm. To ensure the correctness of the Dlog algorithm, it verfiies the potential power and loops until a correct power is found. It turns out that the algorithm converges if the error is less than 50%.

The algorithm is tested for base 3 in mod 5, where the group order is 4 = 22.  The HB function in this case gives no error. The algorithm is also tested for base 3 in mod 17,  for
∣∣𝑍∗17∣∣=16.
Z17∗=16.

 We conducted an experiment that runs in five trials to test the HB function, According to the results of this experiment, the accuracy of the oracle in predicting HB values is 35/80*100 = 56.25 percent, while the error is 43.75 percent. Which implies that the proposed Dlog algorithm converges to a correct value.